when a user quits with the same name as an emote, the emote autocomplete no longer gets removed. bumped version 2.3.2

+++ assets/chat/js/autocomplete.js
@@ -237,10 +237,14 @@ class ChatAutoComplete {
         return data
     }
 
-    remove(str){
+    remove(str, userOnly=false){
         const bucket = this.buckets.get(getBucketId(str))
-        if(bucket && bucket.has(str))
-            bucket.delete(str)
+        if (bucket && bucket.has(str)) {
+            const a = bucket.get(str)
+            if ((userOnly && !a.isemote) || !userOnly) {
+                bucket.delete(str)
+            }
+        }
     }
 
     select(index){
